refactor(cmd): restructure package layout and improve error handling
- Move command files into internal/portal package
- Convert Main() to StartServer() with proper error returns
- Add comprehensive error handling in server startup flow
- Fix typo in SIGQUIT log message
- Simplify main entry point
